% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_singletons.R
\name{add_singletons}
\alias{add_singletons}
\title{Add singletons}
\usage{
add_singletons(dat, use_max = F)
}
\arguments{
\item{dat}{dataframe, result of get sad}
}
\value{
dataframe with some extra singletons
}
\description{
Add singletons
}
