% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{add_singletons_dataset}
\alias{add_singletons_dataset}
\title{Add singletons to a dataset}
\usage{
add_singletons_dataset(dataset, use_max = TRUE)
}
\arguments{
\item{dataset}{dataset, result of \code{load_dataset}, potentially with many sites}

\item{use_max}{passed to \code{add_singletons}: use the high estimate, or no?}
}
\value{
dataset with estimated rare species added to each site
}
\description{
Add singletons to a dataset
}
