% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manips.R
\name{assign_macdb_manips}
\alias{assign_macdb_manips}
\title{Assign ctrl/treatment designation to MACD sites}
\usage{
assign_macdb_manips(all_di, storage_path = here::here("working-data",
  "macdb_data"))
}
\arguments{
\item{all_di}{diversity indicies df}

\item{storage_path}{where the MACD is}
}
\value{
all_di with study and ctrl/treatment
}
\description{
Assign ctrl/treatment designation to MACD sites
}
