% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manips.R
\name{assign_portal_manips}
\alias{assign_portal_manips}
\title{Parse portal site names for manips}
\usage{
assign_portal_manips(all_di)
}
\arguments{
\item{all_di}{all di}
}
\value{
all di with cols for year, plot, season, trtmt
}
\description{
Parse portal site names for manips
}
