% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rov_sampling.R
\name{build_net}
\alias{build_net}
\title{Build net of S and N combinations to sample}
\usage{
build_net()
}
\value{
table of s0, n0, "site name"
}
\description{
Covering a range of values present in the datasets
}
