% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{compare_props_fs}
\alias{compare_props_fs}
\title{Compare PO over whole dataset}
\usage{
compare_props_fs(fs_samples_df, ncomps = 500)
}
\arguments{
\item{fs_samples_df}{fs draws}

\item{ncomps}{comparisons per focal}
}
\value{
comparisons for all draws
}
\description{
Compare PO over whole dataset
}
