% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{compare_props_off_full}
\alias{compare_props_off_full}
\title{Compare proportion off of two sims}
\usage{
compare_props_off_full(focal_sim, fs_df, ncomps = 100)
}
\arguments{
\item{focal_sim}{index of sim}

\item{fs_df}{samples from FS}

\item{ncomps}{how many comparisons}
}
\value{
mean proportion off of focal sim vs ncomparisons sims from fs_df
}
\description{
Compare proportion off of two sims
}
