% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{count_below}
\alias{count_below}
\title{Count values below a value}
\usage{
count_below(a_value, a_vector)
}
\arguments{
\item{a_value}{Focal value}

\item{a_vector}{Vector for comparison}
}
\value{
Number of values in vector below value
}
\description{
Count values below a value
}
