% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_elements.R
\name{count_elements}
\alias{count_elements}
\title{Count number of elements in feasible set}
\usage{
count_elements(s, n, p_table = NULL)
}
\arguments{
\item{s}{S0}

\item{n}{N0}

\item{p_table}{Optionally pass the p table so you don't have to regenerate it}
}
\value{
number of elements in the FS
}
\description{
Count number of elements in feasible set
}
