% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_macdb}
\alias{download_macdb}
\title{Download MACDB data}
\usage{
download_macdb(from_url = FALSE, storage_path = here::here("working-data"))
}
\arguments{
\item{from_url}{defaults FALSE. If true, from figshare. Otherwise, loads data from storage internal to the package}

\item{storage_path}{where to put the data. Defaults to working-data/macdb_data.}
}
\value{
nothing
}
\description{
Download MACDB data
}
