% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{download_portal_plants}
\alias{download_portal_plants}
\title{Download Portal plants}
\usage{
download_portal_plants(storage_path = here::here("working-data",
  "abund_data"))
}
\arguments{
\item{storage_path}{where it goes}
}
\value{
nothing
}
\description{
Download Portal plants
}
