% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_portal_plants_manips}
\alias{download_portal_plants_manips}
\title{Download Portal plants with manipulations}
\usage{
download_portal_plants_manips(storage_path = here::here("working-data",
  "abund_data"))
}
\arguments{
\item{storage_path}{where it goes}
}
\value{
nothing
}
\description{
Download Portal plants with manipulations
}
