% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{filter_fia_small}
\alias{filter_fia_small}
\title{Filter FIA data - tiny communities}
\usage{
filter_fia_small(
  min_s0 = 3,
  max_s0 = 9,
  max_comm = 10000,
  storage_path = here::here("working-data", "abund_data"),
  save = TRUE
)
}
\arguments{
\item{min_s0}{min s0}

\item{max_s0}{max s0}

\item{max_comm}{max ncommunities}

\item{storage_path}{where to put it}

\item{save}{save it?}
}
\value{
nothing
}
\description{
Following White et al 2012, filtering FIA data to sites with > 10 species. This brings it from 103343 sites to 10355.
}
