% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self_similarity_metrics.R
\name{fs_div}
\alias{fs_div}
\title{KL Divergence of two fs}
\usage{
fs_div(two_fs, log = F)
}
\arguments{
\item{two_fs}{two fs}

\item{log}{log units? idk if htis matters}
}
\value{
kl div
}
\description{
KL Divergence of two fs
}
