% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self_similarity_metrics.R
\name{fs_r2}
\alias{fs_r2}
\title{R2 for FS}
\usage{
fs_r2(two_fs, log = F)
}
\arguments{
\item{two_fs}{wide dataframe or matrix: 2 rows, nspp columns}

\item{log}{use log or no? default F}
}
\value{
r2
}
\description{
R2 as calculated in White et al (2012) and others (I beleive macroecotools)
}
