% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{get_percentile}
\alias{get_percentile}
\title{Get one percentile value}
\usage{
get_percentile(a_value, a_vector)
}
\arguments{
\item{a_value}{Focal value}

\item{a_vector}{Comparison vector}
}
\value{
Percentile of focal value within comparison vector
}
\description{
Calculate the percentile value for a focal value relative to a comparison vector, as the \% of elements in the comparison vector less than or equal to the focal value.
}
