% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{get_percentiles}
\alias{get_percentiles}
\title{Get percentile values}
\usage{
get_percentiles(a_vector, incl = T)
}
\arguments{
\item{a_vector}{Vector of values}

\item{incl}{tf lessthan or equal to or just less than}
}
\value{
Vector of percentile values for all values in the vector
}
\description{
Calculate the percentile values (\% of elements in the vector <= a value) for all values in a vector
}
