% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{load_dataset}
\alias{load_dataset}
\title{Load a dataset}
\usage{
load_dataset(
  dataset_name,
  storage_path = here::here("working-data", "abund_data")
)
}
\arguments{
\item{dataset_name}{"bbs", "fia", "gentry", "mcdb", "portal_plants", "misc_abund_short", or "misc_abund"}

\item{storage_path}{where the data is living}
}
\value{
Dataset ready for analysis
}
\description{
Loads a (pre-downloaded, and filtered if relevant) dataset and processes it to the right shape for the rest of the functions.
}
