% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{load_sad}
\alias{load_sad}
\title{Load sads}
\usage{
load_sad(dataset_name, site_name,
  storage_path = here::here("working-data", "paper"))
}
\arguments{
\item{dataset_name}{"bbs", "fia", "gentry", "mcdb", or "misc_abund"}

\item{site_name}{which site}

\item{storage_path}{where the data is living}
}
\value{
something
}
\description{
Load sads
}
