% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{load_sads}
\alias{load_sads}
\title{Load sads}
\usage{
load_sads(storage_path = here::here("working-data", "paper"))
}
\arguments{
\item{storage_path}{}
}
\value{
something
}
\description{
Load sads
}
