% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{prep_macdb}
\alias{prep_macdb}
\title{Prepare MACDB for pipeline}
\usage{
prep_macdb(storage_path = here::here("working-data", "macdb_data"),
  save = TRUE)
}
\arguments{
\item{storage_path}{where to get the data}

\item{save}{defaults TRUE}
}
\value{
nothing
}
\description{
Prepare MACDB for pipeline
}
