% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self_similarity_metrics.R
\name{proportion_off}
\alias{proportion_off}
\title{Proportion off}
\usage{
proportion_off(two_fs)
}
\arguments{
\item{two_fs}{wide dataframe or matrix: 2 rows, nspp columns}
}
\value{
prop diff
}
\description{
Prop of individuals allocated differently.
}
