% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{pull_di}
\alias{pull_di}
\title{Pull observed diversity indices and nb samples achieved}
\usage{
pull_di(di_df)
}
\arguments{
\item{di_df}{result of di_wrapper}
}
\value{
di_df for observed + n samples
}
\description{
Pull observed diversity indices and nb samples achieved
}
