% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self_similarity_metrics.R
\name{rep_diff_sampler}
\alias{rep_diff_sampler}
\title{Repeatedly sample fs set}
\usage{
rep_diff_sampler(fs_set, ndraws)
}
\arguments{
\item{fs_set}{bank of fs samples}

\item{ndraws}{how many reps}
}
\value{
df of comparisons
}
\description{
Repeatedly sample fs set
}
