% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{add_dis}
\alias{add_dis}
\title{Skewness}
\usage{
add_dis(fs_samples_df)
}
\arguments{
\item{fs_samples_df}{df of all samples, with column for abundance}
}
\value{
summary of skewness per sample
}
\description{
Skewness
}
