% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{add_singletons_dataset}
\alias{add_singletons_dataset}
\title{Add singletons to a dataset}
\usage{
add_singletons_dataset(dataset, use_max = TRUE)
}
\arguments{
\item{dataset}{the dataset}

\item{use_max}{use max?}
}
\value{
dataset plus singletons
}
\description{
Add singletons to a dataset
}
