% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{download_data}
\alias{download_data}
\title{Download SAD data}
\usage{
download_data(from_url = FALSE,
  storage_path = here::here("working-data"))
}
\arguments{
\item{from_url}{defaults FALSE. If true, downloads White/Baldridge data directly from GitHub and figshare. Otherwise, loads data from storage internal to the package}

\item{storage_path}{where to put the data. Defaults to working-data/data.}
}
\value{
nothing
}
\description{
Download SAD data
}
