% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{filter_miscabund}
\alias{filter_miscabund}
\title{Filter misc abund data}
\usage{
filter_miscabund(max_s = 200, max_n = 40720,
  storage_path = here::here("working-data", "abund_data"))
}
\arguments{
\item{max_s}{currently 200}

\item{max_n}{currently 40720}

\item{storage_path}{}
}
\value{
nothing
}
\description{
Misc abund dataset has some very large communities, which I am removing for now to keep the p table tractable. This function is to make that filtering documented.
}
