% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{get_percentiles}
\alias{get_percentiles}
\title{Get percentile values}
\usage{
get_percentiles(a_vector)
}
\arguments{
\item{a_vector}{Vector of values}
}
\value{
Vector of percentile values for all values in the vector
}
\description{
Get percentile values
}
