% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{load_dataset}
\alias{load_dataset}
\title{Load a dataset}
\usage{
load_dataset(dataset_name, storage_path = here::here("working-data",
  "paper"))
}
\arguments{
\item{dataset_name}{"bbs", "fia", "gentry", "mcdb", or "misc_abund"}

\item{storage_path}{where the data is living}
}
\value{
something
}
\description{
Load a dataset
}
