% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_reports.R
\name{render_report}
\alias{render_report}
\title{Render a report}
\usage{
render_report(report_path, dependencies = list())
}
\arguments{
\item{report_path}{file path for report}

\item{dependencies}{dependencies for drake}
}
\value{
nothing
}
\description{
Render a report
}
