% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_fs.R
\name{sample_fs_wrapper}
\alias{sample_fs_wrapper}
\title{Sample fs wrapper}
\usage{
sample_fs_wrapper(dataset, site_name, singletonsyn, n_samples,
  p_table = NULL, seed = NULL)
}
\arguments{
\item{dataset}{Full dataset}

\item{site_name}{Which site within the dataset}

\item{singletonsyn}{Singletons?}

\item{n_samples}{nb samples}

\item{p_table}{P table to pass}

\item{seed}{pass a seed for reproducibility}
}
\value{
df of samples from fs
}
\description{
Sample fs wrapper
}
